#41 Update Test for Profile component and rename "util" folder
