+++ oxygen-account-frontend/src/profile/Profile.jsx
@@ -2,7 +2,7 @@ import "../css/GlobalStyle.css";
 import "../css/ProfileStyle.css";
 
 import React from "react";
-import AppHeader from "../util/AppHeader.jsx";
+import AppHeader from "../shared/AppHeader.jsx";
 import ProfileCard from "./ProfileCard.jsx";
 import { Grid } from "@mui/material";
