+++ oxygen-account-frontend/src/shared/OxygenAvatar.jsx
@@ -41,17 +41,15 @@ function OxygenAvatar( {name} ) {
      * @returns {Object} The object for the avatar.
      */
     const stringAvatar = (name) => {
-        return {
-            sx: {
-                bgcolor: stringToColor(name),
-            },
-            children: `${name.split(' ')[0][0] || ''}${name.split(' ')[1] ? name.split(' ')[1][0] : ''}`,
-        };
+        return `${name.split(' ')[0][0] || ''}${name.split(' ')[1] ? name.split(' ')[1][0] : ''}`;
     };
 
     return (
         // Render the Avatar component
-        <Avatar {...stringAvatar(name)} />
+        <Avatar
+            children = {stringAvatar(name)}
+            sx= {{bgcolor: stringToColor(name)}}
+        />
     );
 }
